refactor : enum Color와 enum Name을 Piece 클래스 내부로 이동
- Name을 Type으로 rename합니다.
