--- main/chess/Color.java
@@ -1,5 +1,0 @@
-package chess;
-
-public enum Color {
-    WHITE, BLACK
-}
