--- main/chess/Name.java
@@ -1,5 +1,0 @@
-package chess;
-
-public enum Name {
-    PAWN, KNIGHT, ROOK, BISHOP, QUEEN, KING
-}
